requirements.txt
# please install pip requirements using: 
# cat requirements.txt | sed -e '/^\s*#.*$/d' -e '/^\s*$/d' | xargs -n 1 pip install
# otherwise one single compilation error could eat away like 20 minutes for nothing.
# pip install -e . --config-settings editable_mode=strict --no-build-isolation --no-deps
# pip install -e . --no-build-isolation --no-deps

# other requirements
# for terminal image visualization
timg
tyro
yacs
yapf
tqdm
rich
shtab
sympy
pillow
addict
trimesh
imageio
termcolor
tensorboard
scikit-image
scikit-learn
pytorch_msssim
fast-autocomplete
torch-tb-profiler

# other requirements not available in conda
smplx
pymcubes
torchdiffeq
opencv-python

# dev requirements
h5py

# for plenoctree conversion
# svox # never used
ipdb
pdbr
ninja
lpips
ujson
pandas

# for unwrapping to get StVK properly
xatlas
kornia
msgpack
jupyter
openpyxl
autopep8
pyntcloud
pyturbojpeg
matplotlib
ruamel.yaml
commentjson

# to support visualization, we need:
git+https://github.com/wookayin/gpustat
# for winding_number_remesh
git+https://github.com/NVlabs/nvdiffrast
git+https://github.com/tatsy/torchmcubes
# git+https://github.com/dendenxu/cholespy
# installing from conda leads to just numerous problems...
# NOTE: IMPORTANT
git+https://github.com/facebookresearch/pytorch3d
git+https://github.com/facebookresearch/detectron2
# git+https://github.com/dendenxu/large-steps-pytorch # FIXME: FIX THIS
# git+https://github.com/dendenxu/bvh-ray-tracing # FIXME: FIX THIS
# git+https://github.com/YuliangXiu/bvh-distance-queries

# external dependency: easymocap-public (this repo is not publicly available yet)
# for easymocap's vposer: human_pose_prior, this looks like my DotDict implementation... just way more complex
dotmap
# for easymocap loading of SMPL (maybe all pickle loading of SMPL?)
chumpy
mediapipe
func_timeout
pycocotools
tensorboardX
# git+https://github.com/mmatl/pyopengl # MARK: might not render
git+https://github.com/nghorbani/human_body_prior

# https://storage.googleapis.com/open3d-releases-master/python-wheels/open3d-0.16.0-cp310-cp310-manylinux_2_27_x86_64.whl
# http://www.open3d.org/docs/latest/getting_started.html (install the development version from here if the previsou link is expired and python is too new)
# python3.10 support for open3d finally here
# if failed to install open3d (even when installing from latest release?), try to skip it using
# pip install $(grep -v '^ *#\|^open3d' requirements.txt | grep .)
# NOTE: IMPORTANT
open3d

# pip install $(grep -v '^ *#\|^.*open3d\|^torch-sparse\|^torch-geometric\|^.*cholespy\|^.*pytorch3d\|^.*pyopengl' requirements.txt | grep .)
# cat requirements.txt | sed -e '/^\s*#.*$/d' -e '/^\s*$/d' | xargs -n 1 pip install

# Blender utils
# bpy
memory-tempfile

# dearpygui # TOO SLOW
# imgui[glfw] # replaced with imgui_bundle
glfw
PyGLM
pyperclip
pyopengl
clang-format
imgui-bundle
opencv-contrib-python

# MARK: The torchvision version has a memory leak
# https://github.com/pytorch/vision/issues/4378
# pynvjpeg

# MARK: Requires some compiling and may easily fail, but not needed for most of the implementations
git+https://github.com/NVlabs/tiny-cuda-nn/#subdirectory=bindings/torch

# MARK: prone to change pytorch version, please install these on demand and manually
# spconv-cu118
# functorch
torch-scatter
# torch-sparse
# torch-geometric

# pip install -e . --no-build-isolation --no-deps

# not needed unless you're using ali's implemenetation of VectorQuantize
einops
# not needed unless you're trying to debug torch memory usage using this package directly
pytorch_memlab

# for ply file io
plyfile
# cuda driver for python, so nice for cuda opengl and pytorch interop, no hassle!
cuda-python

# For extracting meshes
pymeshlab

# For encode jpeg
pyturbojpeg

# For gaussian
git+https://gitlab.inria.fr/bkerbl/simple-knn
git+https://github.com/slothfulxtx/diff-gaussian-rasterization
